feat(auth): create useLogin hook
Refactor login functionality into a custom hook to improve code organization and reusability.

The new useLogin hook provides a centralized and efficient way to manage user login, including handling email/password login and Google OAuth login. This approach simplifies the codebase and makes it easier to maintain and extend the login functionality in the future.
